:sparkles: Feat: 스케줄 기록[record](Diary, Activity) 엔티티
<body>
- 스케줄에 대한 개인 기록, 활동 기록 관련 엔티티 및 기능 생성

- 관련 : #230
