--- storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/domains/diary/exception/ActivityImgException.java
@@ -1,10 +1,0 @@
-package com.namo.spring.db.mysql.domains.diary.exception;
-
-import com.namo.spring.core.common.code.BaseErrorCode;
-import com.namo.spring.core.common.exception.GeneralException;
-
-public class ActivityImgException extends GeneralException {
-	public ActivityImgException(BaseErrorCode code) {
-		super(code);
-	}
-}
