--- storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/domains/diary/repository/ActivityUserRepository.java
@@ -1,8 +1,0 @@
-package com.namo.spring.db.mysql.domains.diary.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.namo.spring.db.mysql.domains.diary.entity.ActivityUser;
-
-public interface ActivityUserRepository extends JpaRepository<ActivityUser, Long> {
-}
