+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/domains/record/exception/ActivityException.java
@@ -1,0 +1,10 @@
+package com.namo.spring.db.mysql.domains.record.exception;
+
+import com.namo.spring.core.common.code.BaseErrorCode;
+import com.namo.spring.core.common.exception.GeneralException;
+
+public class ActivityException extends GeneralException {
+	public ActivityException(BaseErrorCode code) {
+		super(code);
+	}
+}
