+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/domains/record/repository/ActivityRepository.java
@@ -1,0 +1,9 @@
+package com.namo.spring.db.mysql.domains.record.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.namo.spring.db.mysql.domains.record.entity.Activity;
+
+public interface ActivityRepository extends JpaRepository<Activity, Long> {
+
+}
