+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/domains/record/repository/DiaryRepository.java
@@ -1,0 +1,8 @@
+package com.namo.spring.db.mysql.domains.record.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.namo.spring.db.mysql.domains.record.entity.Diary;
+
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
+}
