+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/domains/record/service/DiaryImageService.java
@@ -1,0 +1,33 @@
+package com.namo.spring.db.mysql.domains.record.service;
+
+import java.util.Optional;
+
+import org.springframework.transaction.annotation.Transactional;
+
+import com.namo.spring.core.common.annotation.DomainService;
+import com.namo.spring.db.mysql.domains.record.entity.DiaryImage;
+import com.namo.spring.db.mysql.domains.record.repository.DiaryImageRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@DomainService
+@RequiredArgsConstructor
+public class DiaryImageService {
+	private final DiaryImageRepository diaryImageRepository;
+
+	@Transactional
+	public DiaryImage createDiaryImage(DiaryImage diaryImage) {
+		return diaryImageRepository.save(diaryImage);
+	}
+
+	@Transactional(readOnly = true)
+	public Optional<DiaryImage> readDiaryImage(Long diaryImageId) {
+		return diaryImageRepository.findById(diaryImageId);
+	}
+
+	@Transactional
+	public void deleteDiaryImage(Long diaryImageId) {
+		diaryImageRepository.deleteById(diaryImageId);
+	}
+
+}
